Changed expiration field of LoginResult to i64 (#36)
Reviewed-on: https://git.kundeng.us/phoenix/icarus_models/pulls/36
Co-committed-by: phoenix <kundeng00@pm.me>

+++ src/login_result.rs
@@ -9,7 +9,7 @@ pub struct LoginResult {
     pub token: String,
     #[serde(alias = "token_type")]
     pub token_type: String,
-    pub expiration: i32,
+    pub expiration: i64,
 }
 
 impl Default for LoginResult {
